Added Maven into homework_10_proj_101222
