+++ homework_10_proj_101222/src/main/java/app/Main.java
@@ -1,3 +1,5 @@
+package app;
+
 public class Main {
     public static void main(String[] args) {
         RegistrationApp registrationApp = new RegistrationApp();
