+++ homework_10_proj_101222/src/main/java/app/User.java
@@ -1,3 +1,5 @@
+package app;
+
 import java.time.LocalDate;
 
 public class User {
